+++ app/src/main/java/app/sabre/wzsabre/MainBroadcastReceiver.java
@@ -91,6 +91,7 @@ public class MainBroadcastReceiver extends BroadcastReceiver {
         JSONObject s1 = new JSONObject(); s1.put("id", "chp");  s1.put("name", "CHP Live Feed");      sources.put(s1);
         JSONObject s2 = new JSONObject(); s2.put("id", "waze"); s2.put("name", "Waze");               sources.put(s2);
         JSONObject s3 = new JSONObject(); s3.put("id", "lcs");  s3.put("name", "Caltrans Closures");  sources.put(s3);
+        JSONObject s4 = new JSONObject(); s4.put("id", "fire"); s4.put("name", "Wildfires");          sources.put(s4);
         response.put("supported_sources", sources);
         response.put("request_action",  "app.sabre.wzsabre.FETCH_REQUEST");
         response.put("report_action",   "app.sabre.wzsabre.SUBMIT_REPORT");